docs: move coding conventions into contributing.md (#1751)
Move Go coding conventions from the README.md into the CONTRIBUTING.md file which discusses other requirements for code contributed.

I'm in the process of adding a developing document that includes finer details about how to do things like add dependencies, update them, remove them, as part of the previous work switching to Modules (#1634). I noticed during that work that this section is in the projects readme but is primarily listing more things that someone contributing should be thinking about, and not so much someone who is browsing or importing the repository. It makes more sense for this to live in the contributing document.