services/horizon/expingest: Do not start verification when state is already found to be invalid (#1762)
This commit fixes an issue with state verification that can occur when
deploying Horizon in sequence to multiple nodes.

When deploying a fix for invalid state error to multiple Horizon servers
in sequence it's possible that when new code starts running, the old
code that verifies state is still working. The following sequence is
possible:

H1                               H2
(`System.Run()`)                 (post pipeline hook after `verifyState()`)

- Begin()
- UpdateExpStateInvalid(false)
                                 - (autobegin)
                                 - UpdateExpStateInvalid(true)`
                                   (blocked until H1 commits)
...
- Commit()
                                 - (value updated to `true`)
                                 - (autocommit)

When both transactions commit the final value is equal `true` but should
be reset to `false` when deploying Horizon with a fixed ingestion.

After considering many different options I decided to stop running state
verification routine (`verifyState()`) when state is found to be
incorrect. This way the value won't be overwritten during sequential
deploy when the state was found to be incorrect earlier but will
overwrite the value when it happens during deploy.