services/ticker: add docs about which asset in a pair is the base (#1781)
Add docs about which asset in a trade pair is the base, and which is the counter.

To make it less confusing for new comers looking at the API to determine what the values returned by the API mean in the context of a trade pair.

It was confusing to me that the base is on the right-hand-side because the base is usually the currency on the left-hand-side, at least in traditional markets. Some examples of this:
- https://en.wikipedia.org/wiki/Currency_pair
- https://www.investopedia.com/terms/b/basecurrency.asp

We probably can't change the order at this point, and there are probably a good reason we went with this order that I'm not aware of (@tomquisel?), so documenting it at the top of the API docs should alleviate confusion.