Remove EXTENSIONS from initial schema (#1801)
This commit removes `CREATE EXTENSION` from `1_initial_schema.sql`. To
run `CREATE EXTENSION` you need to have superuser permissions in
postgres which is not the case in most deployments (ex. docker image
fails because of this: #140). It looks like it's a result of removing
`latest.sql` file (#1650) that did not contain `CREATE EXTENSION`.