keypair: add MustRandom (#1907)
Add `MustRandom()` that wraps `Random()` and works the same as `MustParse()`, converting any error returned by `Random()` into a `panic`.

Why:

Make it easier to hack on prototypes where keys are needed, and make `Random` functionally symmetrical with the other common entry point `Parse`. While I was spiking a prototype solution for an idea I found `MustParse()` helpful to keep my non-production code and test code thinner. `Random()` can be a helpful function to use in tests and when prototyping ideas to generate temporary keys on application boot, and the addition of the error checking is repetitive and mostly going to be dropping a panic anyway.

There were no tests for the `Random` functions so I added some basic tests.