txnbuild: add GetSourceAccount to all operations (#1908)
Add `GetSourceAccount()` function to the `txnbuild.Operation` interface and to all implementers, surfacing the source account that exists on every implementer.

Why:

I was working with the txnbuild package and was trying to validate that a transaction only contained operations relating to a specific set of accounts and no others. My code validating the accounts impacted by the operation doesn't have any knowledge about the type of operations that might be in the transaction. Every operation has a source account and every future operation is very likely to have a source account as that is how operations can be composed for multiple accounts in a single transaction. So it's very easy to expand the interface of the `txnbuild.Operation` to require it. I think it's important we make it possible for applications to validate the accounts a transaction will directly impact without needing to know about every type of operation. This will help reduce the changes required to an application when new operations are added.

It would be feasible to use the `xdr` package to get the source account for every operation instead of using txnbuild, but given that txnbuild is friendlier and a developer is more likely to be interacting with it already, it makes sense to make this accessible here.

Known limitations:

If we ever have an operation that cannot have a different source account as the transaction this interface will be broken, but given how source accounts are used in operations to group operations relating to multiple accounts and to separate an account surfacing a sequence number and an account the operation is being performed on, I think this is unlikely.