services/horizon: Add ReadTimeout to HTTP server configuration
This commit adds `ReadTimeout` to Horizon HTTP server configuration.

It's possible to send a POST request with `Content-Length` header with a
large value and then send the request body very slowly (ex. 1 byte per
second). When many connections like this are established it's possible
that the server will run out of file descriptors to serve legitimate
users.

To solve this we limit the maximum duration for reading the entire
request, including the body to 5 seconds.

It shouldn't affect any legitimate `GET` requests because the
`ReadHeaderTimeout` has been set since Horizon 0.15.0. For `POST
/transactions` request after this commit entire body must be written to
Horizon in 5 seconds. It seems that it's enough time to send any
transaction envelope.