services/keystore: update changelog to signal current revision as v1.1.1 (#1959)
Update the Keystore CHANGELOG to signal that the current revision is v1.2.0.

Also, I re-ordered the versions so they are in descending order, like our other changelogs.

Why:
An important fix was made to Keystore and we should signal that the current version is a usable version of Keystore.

Known limitations:
It appears we haven't been tagging releases or releasing binaries of Keystore at the moment because our only deployment of it is continuously deployed commit-by-commit irrespective of versions. I think given the two fixes are noteworthy it's worth signaling them as v1.2.0 to encourage someone who might be happen to checking out the application to use a new version oft he application but I don't want to address the tagging or releasing in this release especially as this service is still really experimental and we're not using it as we had intended.