exp/pipeline: Graceful shutdown of Pipeline (#1972)
When `Pipeline.Shutdown()` was called post-processing hooks were not
triggered. This could lead to a DB connection leak. To fix this,
pipeline code now sends `ErrShutdown` if the pipeline was shutdown in
`Process` (if shutdown happened during or before processing) and to
post-processing hooks. It also always sends post-processing hooks (even
when pipeline was shutdown).

Next, `App.CloseDB()` is called when `expingest.System.Run()` returns
what means that all the DB connections can be properly closed without a
possibility of triggering `database closed` errors.

Finally, ledger and state reporters were updated with proper log
messages when pipeline was shutdown.