all: Use context from session in sqlx functions (#1977)
Use the context in `db.Session` when executing sql queries and
statements on the session connection. A consequence of this change is
that the session context is no longer optional.

It should be possible to pass context.Context to set timeout for queries
sent to a database. This is to prevent long running queries to occupy DB
connection for a long period of time.