services/horizon: Fix ledger processor when not ingesting into a database (#1992)
All processors updating a database are ancestors of `ContextFilter` with
`IngestUpdateDatabase` set. `ContextFilter` will only pass transactions
to the stream writer if the given value in the context is present,
otherwise it will close the writer (but the processors below it, will
still be executed . This works well for existing processors but ledgers
processor is different: instead of operating on per transactions basis,
it inserts data into a DB (ledger header) only once.

The previous code was trying to insert ledger header into a DB even when
it shouldn't (`IngestUpdateDatabase` not present in the context).