support/config: simplify initing and setting multiple ConfigOptions (#2042)
Adds a `ConfigOptions` type that is a slice of `ConfigOptions`, and wraps up initializing and setting the values of all the options. Also updates Horizon to use those new functions.

Why:
I had a go of using the `ConfigOption` type and following the same pattern in Horizon in some new code and I lost a lot of time to a silly mistake when defining the slice of options that caused none of the options to ever be set (I forgot to use a pointer type). I think this code will help prevent that, but also this code reduces the repetitiveness of using this code in another application.

I think it's great that we have some code to help us setup flags and environment variables in a common format. I think this is one small thing we can do to improve this code.