support/http: remove behind proxy option for muxes (#2051)
Remove the ability to configure `support/http` muxes/routers with the behind proxy flag.

We use these mux functions in five places across four applications, but always with behind proxy as false. We never offer a way to configure it, it's always hardcoded. The middleware that behind proxy adds is the chi `RealIP` middleware, and even in Horizon we don't use it. It seems like something we added for a specific use case but never used. Removing it to make this code simpler.

This is a non-breaking change for the applications in the repo because I'm updating all the callers to not pass the value, but this is a breaking change for anyone outside this repository who might be importing `github.com/stellar/go/support/http`. As I understand it the code that lives under the `support` package is primarily intended for use by services within this repository and so the likelihood of this being a breaking change for anyone is small. According to [GoDoc.org there are no importers](https://godoc.org/github.com/stellar/go/support/http?importers) of `support/http` who aren't a fork of this repository.

We'd benefit from a formal backwards compatibility statement that classified each package and what its compatibility promise is. If we ever decide one I think we should keep the support package outside any compatibility promise